requirements.txt
dj_database_url==1.2.0
Django==4.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.2.2
PyJWT==2.8.0
python-dotenv==1.0.1
rest_framework_simplejwt==0.0.2
validators==0.28.1
